Moved som classes to a seperate file
